fix: agregar funciones faltantes en tenant_service y modelos relacionados
- obtener_siguiente_numero_cotizacion y obtener_siguiente_numero_ot en tenant_service
- Modelos, schemas y routers actualizados para cotizaciones y OT
- Fix ImportError que impedía arrancar el backend en producción

+++ schemas/movimiento_inventario.py
@@ -29,7 +29,7 @@ class MovimientoInventarioResponse(BaseModel):
     producto_id: int
     local_origen_id: Optional[int]
     local_destino_id: Optional[int]
-    cantidad: int
+    cantidad: float
     tipo_movimiento: str
     referencia_id: Optional[int]
     notas: Optional[str]
